tests/gem_largeobject: silence the compiler
gem_largeobject.c: In function ‘test_large_object’:
gem_largeobject.c:95:20: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]

+++ tests/gem_largeobject.c
@@ -92,7 +92,7 @@ test_large_object(int fd)
 
 	pwrite.handle = create.handle;
 	pwrite.size = obj_size;
-	pwrite.data_ptr = (uint64_t)data;
+	pwrite.data_ptr = (uintptr_t)data;
 
 	ret = ioctl(fd, DRM_IOCTL_I915_GEM_PWRITE, &pwrite);
 	if (ret) {
